+++ tests/test_clients.py
@@ -153,6 +153,7 @@ class BaseClientTest(unittest.TestCase):
         self.assertEqual(self.base_client.trading_controls, [])
         self.assertTrue(self.base_client.order_stream)
         self.assertIsNone(self.base_client.order_stream_conflate_ms)
+        self.assertEqual(self.base_client.order_streaming_timeout, 0.25)
         self.assertTrue(self.base_client.best_price_execution)
         self.assertTrue(self.base_client.min_bet_validation)
         self.assertFalse(self.base_client.paper_trade)