+++ tests/test_streams.py
@@ -36,7 +36,9 @@ class StreamsTest(unittest.TestCase):
         mock_client.VENUE = streams.VenueType.BETFAIR
         self.streams.add_client(mock_client)
         add_betfair_order_stream.assert_called_with(
-            mock_client, conflate_ms=mock_client.order_stream_conflate_ms
+            mock_client,
+            conflate_ms=mock_client.order_stream_conflate_ms,
+            streaming_timeout=mock_client.order_streaming_timeout,
         )
 
     @mock.patch("flumine.streams.streams.Streams.add_simulated_order_stream")
@@ -46,7 +48,9 @@ class StreamsTest(unittest.TestCase):
         )
         mock_client.VENUE = streams.VenueType.BETFAIR
         self.streams.add_client(mock_client)
-        mock_add_simulated_order_stream.assert_called_with(mock_client)
+        mock_add_simulated_order_stream.assert_called_with(
+            mock_client, streaming_timeout=mock_client.order_streaming_timeout
+        )
 
     @mock.patch("flumine.streams.streams.Streams.add_betfair_order_stream")
     def test_add_client_no_order_stream(self, mock_add_betfair_order_stream):
@@ -62,7 +66,9 @@ class StreamsTest(unittest.TestCase):
         )
         mock_client.VENUE = streams.VenueType.BETDAQ
         self.streams.add_client(mock_client)
-        mock_add_order_stream.assert_called_with(mock_client)
+        mock_add_order_stream.assert_called_with(
+            mock_client, streaming_timeout=mock_client.order_streaming_timeout
+        )
 
     @mock.patch("flumine.streams.streams.Streams.add_stream")
     def test_add_client_order_stream(self, mock_add_stream):